requirements.txt
MySQL-python==1.2.3
PyYAML==3.11
